triton_ce executor: don't error if triton is not installed
